+++ src/actors/message.rs
@@ -5,6 +5,7 @@ use crate::types::{Metadata, Timestamps};
 pub type VideoId = String;
 pub type VideoTitle = String;
 
+#[derive(Debug)]
 pub struct DownloadedStream {
     pub video_id: String,
     pub file: NamedTempFile,